Move codegen to separate module
