+++ src/codegen/CMakeLists.txt
@@ -1,0 +1,20 @@
+include_directories(
+        ${FAASM_INCLUDE_DIR}/codegen
+        ${FAASM_WAVM_SOURCE_DIR}/Lib
+)
+
+set(HEADERS
+    "${FAASM_INCLUDE_DIR}/codegen/MachineCodeGenerator.h"
+    )
+
+set(LIB_FILES
+    MachineCodeGenerator.cpp
+    ${HEADERS}
+    )
+
+faasm_private_lib(codegen "${LIB_FILES}")
+
+target_link_libraries(codegen
+    wavmmodule
+    wamrmodule
+)
