+++ src/runner/codegen_shared_obj.cpp
@@ -1,16 +1,18 @@
 #include <boost/filesystem.hpp>
+
 #include <faabric/util/environment.h>
 #include <faabric/util/locks.h>
 #include <faabric/util/logging.h>
 #include <faabric/util/string_tools.h>
-#include <storage/FileLoader.h>
+
+#include <codegen/MachineCodeGenerator.h>
 
 using namespace boost::filesystem;
 
 void codegenForDirectory(std::string& inputPath)
 {
     SPDLOG_INFO("Running codegen on directory {}", inputPath);
-    storage::FileLoader& loader = storage::getFileLoader();
+    codegen::MachineCodeGenerator& gen = codegen::getMachineCodeGenerator();
 
     // Iterate through the directory
     path inputFilePath(inputPath);
@@ -22,7 +24,7 @@ void codegenForDirectory(std::string& inputPath)
     std::vector<std::thread> threads;
 
     for (unsigned int i = 0; i < nThreads; i++) {
-        threads.emplace_back([&iter, &mx, &end, &loader] {
+        threads.emplace_back([&iter, &mx, &end, &gen] {
             SPDLOG_INFO("Spawning codegen thread");
 
             while (true) {
@@ -48,7 +50,7 @@ void codegenForDirectory(std::string& inputPath)
                 if (faabric::util::endsWith(fileName, ".so") ||
                     faabric::util::endsWith(fileName, ".wasm")) {
                     SPDLOG_INFO("Generating machine code for {}", thisPath);
-                    loader.codegenForSharedObject(thisPath);
+                    gen.codegenForSharedObject(thisPath);
                 }
             }
         });
@@ -74,7 +76,7 @@ int main(int argc, char* argv[])
     if (is_directory(inputPath)) {
         codegenForDirectory(inputPath);
     } else {
-        storage::FileLoader& loader = storage::getFileLoader();
-        loader.codegenForSharedObject(inputPath);
+        codegen::MachineCodeGenerator& gen = codegen::getMachineCodeGenerator();
+        gen.codegenForSharedObject(inputPath);
     }
 }
