+++ src/storage/FileLoader.cpp
@@ -1,12 +1,5 @@
 #include <conf/FaasmConfig.h>
 #include <storage/FileLoader.h>
-#include <wavm/WAVMWasmModule.h>
-
-#if (WAMR_EXECUTION_MODE_INTERP)
-// Import for codegen not needed as it's not supported
-#else
-#include <wamr/WAMRWasmModule.h>
-#endif
 
 #include <stdexcept>
 
@@ -18,7 +11,6 @@
 #include <boost/algorithm/string.hpp>
 #include <boost/filesystem.hpp>
 #include <boost/filesystem/operations.hpp>
-#include <openssl/md5.h>
 
 using namespace faabric::util;
 
@@ -192,15 +184,6 @@ std::string FileLoader::getHashFilePath(const std::string& path)
     return path + HASH_EXT;
 }
 
-std::vector<uint8_t> FileLoader::hashBytes(const std::vector<uint8_t>& bytes)
-{
-    std::vector<uint8_t> result(MD5_DIGEST_LENGTH);
-    MD5(reinterpret_cast<const unsigned char*>(bytes.data()),
-        bytes.size(),
-        result.data());
-
-    return result;
-}
 
 std::vector<uint8_t> FileLoader::loadHashFileBytes(
   const std::string& path,
@@ -244,9 +227,6 @@ void FileLoader::uploadFunction(faabric::Message& msg)
     // Note, when uploading, the input data is the function body
     const std::string& inputBytes = msg.inputdata();
     uploadFileString(key, localCachePath, inputBytes);
-
-    // Build the object file from the file we've just received
-    codegenForFunction(msg);
 }
 
 // -------------------------------------
@@ -451,104 +431,6 @@ void FileLoader::uploadSharedFile(const std::string& path,
 }
 
 // -------------------------------------
-// CODEGEN
-// -------------------------------------
-
-std::vector<uint8_t> FileLoader::doCodegen(std::vector<uint8_t>& bytes,
-                                           const std::string& fileName,
-                                           bool isSgx)
-{
-    if (conf.wasmVm == "wamr") {
-#if (WAMR_EXECTION_MODE_INTERP)
-        throw std::runtime_error(
-          "WAMR codegen not supported with WAMR interp mode");
-#else
-        return wasm::wamrCodegen(bytes, isSgx);
-#endif
-    } else {
-        assert(isSgx == false);
-        return wasm::wavmCodegen(bytes, fileName);
-    }
-}
-
-void FileLoader::codegenForFunction(faabric::Message& msg)
-{
-    std::vector<uint8_t> bytes = loadFunctionWasm(msg);
-
-    const std::string funcStr = funcToString(msg, false);
-
-    if (bytes.empty()) {
-        throw std::runtime_error("Loaded empty bytes for " + funcStr);
-    }
-
-    // Compare hashes
-    std::vector<uint8_t> newHash = hashBytes(bytes);
-    std::vector<uint8_t> oldHash;
-    if (conf.wasmVm == "wamr") {
-        oldHash = loadFunctionWamrAotHash(msg);
-    } else if (conf.wasmVm == "wavm" && msg.issgx()) {
-        SPDLOG_ERROR("Can't run SGX codegen for WAVM. Only WAMR is supported.");
-        throw std::runtime_error("SGX codegen for WAVM");
-    } else {
-        oldHash = loadFunctionObjectHash(msg);
-    }
-
-    if ((!oldHash.empty()) && newHash == oldHash) {
-        SPDLOG_DEBUG("Skipping codegen for {}", funcStr);
-        return;
-    } else if (oldHash.empty()) {
-        SPDLOG_DEBUG("No old hash found for {}", funcStr);
-    } else {
-        SPDLOG_DEBUG("Hashes differ for {}", funcStr);
-    }
-
-    // Run the actual codegen
-    std::vector<uint8_t> objBytes;
-    try {
-        objBytes = doCodegen(bytes, funcStr, msg.issgx());
-    } catch (std::runtime_error& ex) {
-        SPDLOG_ERROR("Codegen failed for " + funcStr);
-        throw ex;
-    }
-
-    // Upload the file contents and the hash
-    if (conf.wasmVm == "wamr") {
-        uploadFunctionWamrAotFile(msg, objBytes);
-        uploadFunctionWamrAotHash(msg, newHash);
-    } else {
-        uploadFunctionObjectFile(msg, objBytes);
-        uploadFunctionObjectHash(msg, newHash);
-    }
-}
-
-void FileLoader::codegenForSharedObject(const std::string& inputPath)
-{
-    // Load the wasm
-    std::vector<uint8_t> bytes = loadSharedObjectWasm(inputPath);
-
-    // Check the hash
-    std::vector<uint8_t> newHash = hashBytes(bytes);
-    std::vector<uint8_t> oldHash = loadSharedObjectObjectHash(inputPath);
-
-    if ((!oldHash.empty()) && newHash == oldHash) {
-        SPDLOG_DEBUG("Skipping codegen for {}", inputPath);
-        return;
-    }
-
-    // Run the actual codegen
-    std::vector<uint8_t> objBytes = doCodegen(bytes, inputPath);
-
-    // Do the upload
-    if (conf.wasmVm == "wamr") {
-        throw std::runtime_error(
-          "Codegen for shared objects not supported with WAMR");
-    }
-
-    uploadSharedObjectObjectFile(inputPath, objBytes);
-    uploadSharedObjectObjectHash(inputPath, newHash);
-}
-
-// -------------------------------------
 // PYTHON FUNCTIONS
 // -------------------------------------
 
@@ -612,7 +494,7 @@ void FileLoader::uploadPythonFunction(faabric::Message& msg)
     uploadFileString(key, "", inputBytes);
 }
 
-void convertMessageToPython(faabric::Message& msg)
+void FileLoader::convertMessageToPython(faabric::Message& msg)
 {
     msg.set_ispython(true);
     msg.set_pythonfunction(msg.function());
