+++ src/upload/CMakeLists.txt
@@ -10,6 +10,7 @@ faasm_private_lib(upload_lib "${LIB_FILES}")
 target_link_libraries(upload_lib
         wasm
         wavmmodule
+        codegen
         faabric
         cpprestsdk::cpprest
         )
